+++ services/Skyra.Database/Models/PointsQuery.cs
@@ -2,7 +2,7 @@
 {
 	public struct PointsQuery
 	{
-		public long Points { get; set; }
+		public int Points { get; set; }
 		public bool Success { get; set; }
 		public string? FailureReason { get; set; }
 	}
